Add PR feedback, need to remove test code and add unit tests
